review: move planData conversion function to planData module
